Time: 676 ms (5.13%), Space: 78.5 MB (70.77%) - LeetHub

+++ 853-car-fleet/853-car-fleet.cpp
@@ -7,11 +7,11 @@ public:
             arr.push_back({position[idx], speed[idx]});
         }
         sort(arr.begin(), arr.end());
-        stack<pair<double,double>> sk;
+        stack<pair<int,int>> sk;
         sk.push(arr[0]);
         for (int idx = 1; idx < n; ++idx) {
             while(!sk.empty() && (sk.top().second > arr[idx].second) &&
-                  (sk.top().first-arr[idx].first)/(arr[idx].second-sk.top().second)*arr[idx].second+arr[idx].first <= target) {
+                  (double)(sk.top().first-arr[idx].first)/(arr[idx].second-sk.top().second)*arr[idx].second+arr[idx].first <= target) {
                 sk.pop();
             }
             sk.push(arr[idx]);
